requirements.txt
django-bootstrap-icons
django-crispy-forms
crispy-bootstrap5
django-bootstrap-v5
python-ldap
django_auth_ldap
django-filter
Pillow
psycopg2-binary
wand